Read servers also from servers_config_local.json
Because previously I had to modify `servers_config.json` and then that
git-managed file shows up as being modified and it's easy to
accidentally commit it and even if I don't commit it, it might make
switching branches or rebasing harder because it's modified.

Solution is to also load from a `servers_config_local.json` file which
is not committed to git (in fact, it's in `.gitignore`) so I can do all
my local customizations there without messing up git.